backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
redis==5.0.8
rq==2.6.0
pydantic==2.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
bcrypt==3.2.2
psycopg2==2.9.9
cryptography
apscheduler
numpy
pillow
pytest_asyncio
httpx